TightlyBound.jl Documentation

A three-body tight binding program written in Julia

Primary Author: Kevin F. Garrity, NIST

Note

This package currently barely works. Needs much more testing.

Package Features

Run tight-binding calculations with near DFT level accuracy (PBEsol).

Get results in seconds based on pre-fit parameters from across periodic table.

Calculate bandstructures and total energies.

Get forces, stresses, and relax structures.

Parameters based on two- and three-body interactions.

Includes self-consistent treatment of long-range Coulomb interaction.

Plotting based on interface of Plots.jl

Outline

Stuff

Other stuff

Index

TightlyBound.TightlyBound

TightlyBound.CrystalMod.makecrys

TightlyBound.relax_structure

TightlyBound.TightlyBound — Module

holds three body tight binding important stuff

TightlyBound.CrystalMod.makecrys — Function

makecrys(A,coords,types)

Return a crystal object from 3×3 lattice A (Bohr), nat × 3 coords in crystal units, and nat element strings

Note: also export-ed directly from TightlyBound for convenience

#Examples

julia> makecrys([10.0 0 0; 0 10.0 0; 0 0 10.0], [0.0 0.0 0.0], ["H"])
A1=     10.00000  0.00000  0.00000
A2=     0.00000  10.00000  0.00000
A3=     0.00000  0.00000  10.00000

H    0.00000  0.00000  0.00000

makecrys(filename::String)

Read filename, return crystal object. File can be POSCAR, or simple quantum espresso inputfile

makecrys(lines::Array{String,1})

Read string array, return crystal object. string array can be POSCAR, or simple quantum espresso inputfile

TightlyBound.relax_structure — Function

relax_structure(c::crystal; mode="vc-relax")

Find the lowest energy atomic configuration of crystal c.

...

Arguments

c::crystal: the structure to relax, only required argument

mode="vc-relax": Default (variable-cell relax) will relax structure and cell, anything else will relax structure only.

database=missing: coefficent database, default is to use the pre-fit pbesol database

smearing=0.01: smearing temperature (ryd), default = 0.01

grid=missing: k-point grid, e.g. [10,10,10], default chosen automatically

nsteps=100: maximum iterations

update_grid=true: update automatic k-point grid during relaxation

...