TightlyBound

TightlyBound.TightlyBound — Module

holds three body tight binding important stuff

TightlyBound.relax_structure — Method

relax_structure(c::crystal; mode="vc-relax")

Find the lowest energy atomic configuration of crystal c.

Arguments

c::crystal: the structure to relax, only required argument

mode="vc-relax": Default (variable-cell relax) will relax structure and cell, anything else will relax structure only.

database=missing: coefficent database, default is to use the pre-fit pbesol database

smearing=0.01: smearing temperature (ryd), default = 0.01

grid=missing: k-point grid, e.g. [10,10,10], default chosen automatically

nsteps=100: maximum iterations

update_grid=true: update automatic k-point grid during relaxation

TightlyBound.scf_energy — Method

scf_energy(c::crystal)

Calculate energy, force, and stress for a crystal. Does self-consistent-field (SCF) calculation if using self-consistent electrostatics.

returns energy, tight-binding-crystal-object, error-flag

Arguments

c::crystal: the structure to calculate on. Only required argument.

database=missing: Source of coeficients. Will be loaded from pre-fit coefficients if missing.

smearing=0.01: Gaussian smearing temperature, in Ryd. Usually can leave as default.

grid=missing: k-point grid, e.g. [10,10,10], default chosen automatically

conv_thr = 1e-5: SCF convergence threshold (Ryd).

iter = 75: number of iterations before switch to more conservative settings.

mix = -1.0: initial mixing. -1.0 means use default mixing. Will automagically adjust mixing if SCF is failing to converge.

mixing_mode =:pulay: default is Pulay mixing (DIIS). Other option is :simple, for simple linear mixing of old and new electron-density. Will automatically switch to simple if Pulay fails.

TightlyBound.scf_energy — Method

scf_energy(tbc::tbc_crys)

SCF energy using crystal structure from TBC object.

TightlyBound.scf_energy — Method

scf_energy(d::dftout)

SCF energy using crystal structure from DFT object.

TightlyBound.scf_energy_force_stress — Method

scf_energy_force_stress(c::crystal; database = missing, smearing = 0.01, grid = missing)

Calculate energy, force, and stress for a crystal.

returns energy, force, stress, tightbindingcrystal_object

Arguments

c::crystal: the structure to calculate on. Only required argument.

database=missing: Source of coeficients. Will be loaded from pre-fit coefficients if missing.

smearing=0.01: Gaussian smearing temperature, in Ryd. Usually can leave as default.

grid=missing: k-point grid, e.g. [10,10,10], default chosen automatically

TightlyBound.scf_energy_force_stress — Method

scf_energy_force_stress(tbc::tb_crys; database = missing, smearing = 0.01, grid = missing)

Calculate energy, force, and stress for a tight binding crystal object. This allows the calculation to run without re-doing the SCF calculation. Assumes SCF already done!

returns energy, force, stress, tightbindingcrystal_object

TightlyBound.set_units — Method

function set_units(;energy=missing, length=missing, both=missing)

Set global units for energy/length. Run with no arguments to check/return current units.

Default units are "eV" and "Angstrom" (or "Å" or "Ang" ).

Choose atomic units by setting energy="Ryd" and length="Bohr".

Set both at the same time with both="atomic" or both="eVAng"

Internally, all units are atomic. Only main public facing functions actually change units.

TB

TightlyBound.TB.tb — Type

mutable struct tb{T}

Holds key tight-binding information in real-space. Like hrdat file from Wannier90. Also part of the `tbcrys` object.

Holds

H::Array{Complex{T},3} Hamiltonian. nw×nw×nr

ind_array::Array{Int64,3} nr×3 , holds the r-space supercells of the TB object.

r_dict::Dict keys are three Ints like [0,0,0], returns the corresponding ind_array index.

nwan::Int Number of orbitals (generalized wannier functions).

nr::Int64 number of R-space supercells.

nonorth::Bool : true if non-orthogonal. Almost always true in this code.

S::Array{Complex{T},3} : Overlap matrix, organized like H

scf::Bool equal to true if requires self-consistency (usually true for fit tb, false for direct from DFT)

h1::Array{T,2} Has the term determined by scf calculations, if calculated already.

TightlyBound.TB.tb_crys — Type

mutable struct tb_crys{T}

Main tight-binding object, holds the tight-biding model tb and information about the crystal

Holds

tb::tb Has the key tb info (see above)

crys::crystal Has the crystal structure

nelec::Float64 Number of electrons

dftenergy::Float64 DFT energy for reference, only for fit to DFT cases.

scf::Bool true if requires self-consistency.

gamma::Array{T, 2} has the ewald calculation results, needed for self-consistency.

eden::Array{Float64,1} electron density, by orbital, if calculated by self-consistency.

within_fit::Bool is true if model is passes tests of being within the fitting parameter space, false for extrapolation

energy::Float64 energy in Ryd, if calculated.

efermi::Float64 Fermi energy in Ryd, if calculated.

TightlyBound.TB.tb_crys_kspace — Type

mutable struct tb_crys_kspace{T}

Hold k-point tight binding and crystal structure. Similar to tb_crys

Holds

tb::tb_k

crys::crystal

nelec::Float64

dftenergy::Float64

scf::Bool

gamma::Array{T, 2}

eden::Array{Float64,1}

TightlyBound.TB.tb_k — Type

mutable struct tb_k{T}

Tight binding object in k-space. Can be from direct import of DFT band structure using atomic proj, or from fft'ed tb object. Similar to real-space version.

Holds

Hk::Array{Complex{T},3} Hamiltonian in k-space

K::Array{Float64,2} K point array. In fractional coordinates of BZ.

kweights::Array{Float64,1} K point weights.

k_dict::Dict Dictionary from k-point like [0,0,0] to index.

nwan::Int64 Number of orbitals / generalized wannier functions.

nk::Int64 Number of k-points.

nonorth::Bool

Sk::Array{Complex{T},3} Overlap matrix.

scf::Bool needs self-consistency?

h1::Array{T,2} #scf term holds scf term if present

grid::Array{Int64,1} dimensions of k-point grid, from regular grid like [8,8,8]

TightlyBound.TB.Hk — Method

function Hk(hk,sk, h::tb, kpoint)

Hk function with pre-allocated memory hk, sk

TightlyBound.TB.Hk — Method

function Hk(h::tb, kpoint)

Calculate band structure at a k-point from tb

TightlyBound.TB.Hk — Method

function Hk(h::tb_crys, kpoint)

Calculate band structure at a k-point from tb_crys

TightlyBound.TB.Hk — Method

function Hk(h::tb_crys_kspace, kpoint)

Calculate band structure at a k-point from a tb_crys_kspace object. Note, can only return precalculated k-points. Need real-space version to get arbitrary k-points.

#Returns

vect - Eigenvectors numwan × numwan complex matrix at kpoint

vals - Eigenvalues (num_wan)

hk - Hamiltonian at kpoint

sk - Overlap matrix at kpoint

vals0 - <vect | Hk0 | vect> where Hk0 is the non-scf part of the Hamiltonian.

#Arguments

h::tb_crys_kspace - tbcryskspace object

kpoint - e.g. [0.0,0.0,0.0]

scf=missing - default is to take SCF from h.

TightlyBound.TB.Hk — Method

function Hk(h::tb_k, kpoint)

Calculate band structure at a k-point from tb_k. Must be pre-calculated k-point.

TightlyBound.TB.calc_bands — Method

function calc_bands(h, kpoints::Array{Float64,2})

Calculate bandstructure for k-points from k-point array. h is a tb or tb_k object.

TightlyBound.TB.calc_bands — Method

function calc_bands(tbc::tb_crys, kpoints::Array{Float64,2})

Calculate bandstructure for k-points from k-point array. Returns eigenvalues.

Arguments

tbc::tb_crys - The tight binding object

kpoints::Array{Float64,2} - k-point array. e.g. [0.0 0.0 0.0; 0.0 0.0 0.1]

TightlyBound.TB.calc_energy — Method

function calc_energy(h::tb_crys, kgrid; smearing=0.01, returnk=false)

Calculate energy no fft

TightlyBound.TB.calc_energy — Method

function calc_energy(tbc::tb_crys; smearing=0.01, returnk=false)

Calculate energy without fft.

TightlyBound.TB.calc_energy_band — Method

function calc_energy_band(h::tb, nelec, kgrid; smearing=0.01, returnk=false)

calculate energy no fft

TightlyBound.TB.calc_energy_charge_fft — Method

function calc_energy_charge_fft(tbc::tb_crys; grid=missing, smearing=0.01)

Do fft, then calculate energy and charge.

TightlyBound.TB.calc_energy_charge_fft_band — Method

function calc_energy_charge_fft_band(hk3, sk3, nelec; smearing=0.01, h1 = missing)

Calculate energy and charge density. For internal use.

return energy0, efermi, chargeden[:], VECTS, VALS, error_flag

TightlyBound.TB.calc_energy_fft — Method

function calc_energy_fft(tbc::tb_crys; grid=missing, smearing=0.01, return_more_info=false)

Get energy using fft.

returns energy

if return_more_info==true then returns etot, efermi, vals, vects

TightlyBound.TB.calc_energy_fft_band — Method

function calc_energy_fft_band(hk3, sk3, nelec; smearing=0.01, return_more_info=false, h1 = missing)

Return energy from hamiltonian hk3, overlap sk3, nelec, etc. Primarly for internal calling after fft.

TightlyBound.TB.ewald_energy — Function

function ewald_energy(tbc::tb_crys_kspace, delta_q=missing)

TightlyBound.TB.ewald_energy — Function

function ewald_energy(tbc::tb_crys, delta_q=missing)

Return ewald energy term from tbc. If delta_q, the atomic charge density, is missing, loads from tbc.

TightlyBound.TB.ewald_energy — Method

function ewald_energy(crys::crystal, gamma, delta_q::Array{Float64,1})

Does the actual calculation.

TightlyBound.TB.find_vbm_cbm — Method

function find_vbm_cbm(eigs, fermi)

Find the valence band max and conduction band minimum from eigs, relative to Fermi level.

TightlyBound.TB.get_dq — Method

function get_dq(crys::crystal, chargeden::Array{Float64,1})

TightlyBound.TB.get_dq — Method

function get_dq(tbc::tb_crys_kspace)

Get atomic charge density from tb_crys or tb_crys_kspace or crys + eden

TightlyBound.TB.get_dq — Method

function get_dq(tbc::tb_crys)

TightlyBound.TB.get_energy_electron_density_kspace — Method

function get_energy_electron_density_kspace(tbcK::tb_crys_kspace; smearing = 0.01)

Get energy / charge density from k-space tight binding object.

return bandenergy + etypes + echarge + energy_smear, eden, VECTS, VALS, error_flag

TightlyBound.TB.get_energy_electron_density_kspace — Method

function get_energy_electron_density_kspace(tb_k::tb_k, nelec; smearing = 0.01)

K-space get energy and electron density from tb_k

TightlyBound.TB.get_h1 — Method

function get_h1(tbc::tb_crys, chargeden::Array{Float64,1})

TightlyBound.TB.get_h1 — Method

function get_h1(tbc::tb_crys_kspace, chargeden::Array{Float64,1})

TightlyBound.TB.get_h1 — Method

function get_h1(tbc::tb_crys_kspace)

TightlyBound.TB.get_h1 — Method

function get_h1(tbc::tb_crys)

Get H1, the potential term added to tight binding in SCF calculation.

TightlyBound.TB.get_neutral_eden — Function

function get_neutral_eden(crys::crystal, nwan=missing)

TightlyBound.TB.get_neutral_eden — Method

function get_neutral_eden(tbc::tb_crys)

Gets a neutral charge density (no charge transfer) to start SCF calculation.

TightlyBound.TB.get_sym_R — Function

function get_sym_R(crys, grid, sss = 1.0)

Figures out the r-space grid using Wigner-Seitz like construction to figure out the best arrangement of r-grid points to keep periodic copies closest to the original atom and take into account symmetry.

return R_grid, R_int_grid, sym_R

returns the R_grid, the integer version, and the symmetry factor of each point.

TightlyBound.TB.load_hr_dat — Function

function load_hr_dat(filename, directory="")

Load a wannier90 hr.dat file Not curretnly a major part of program, but you can use if you want.

TightlyBound.TB.make_kgrid — Method

function make_kgrid(kgrid)

-kgrid is an array of 3 integers like [8,8,8]

returns regular zero-centered k-point grid and (equal) k-weights.

TightlyBound.TB.make_tb — Method

function make_tb(H, ind_arr, S; h1=missing)

Constructor function for tb, better programming.

TightlyBound.TB.make_tb — Method

function make_tb(H, ind_arr, r_dict::Dict, S; h1=missing)

Constructor function for tb with overlaps

TightlyBound.TB.make_tb — Method

function make_tb(H, ind_arr, r_dict::Dict; h1=missing)

Constructor function for tb

TightlyBound.TB.make_tb_crys — Method

function make_tb_crys(ham::tb,crys::crystal, nelec::Float64, dftenergy::Float64; scf=false, eden = missing, gamma=missing, within_fit=true, screening=1.0, tb_energy=-999, fermi_energy=0.0 )

Constructor function for tb_crys object

TightlyBound.TB.make_tb_crys_kspace — Method

function make_tb_crys_kspace(hamk::tb_k,crys::crystal, nelec::Float64, dftenergy::Float64; scf=false, eden = missing, gamma=missing, screening=1.0)

Constructor function for tb_crys_kspace object

TightlyBound.TB.make_tb_k — Method

function make_tb_k(Hk, K, kweights, Sk; h1=missing, grid=[0,0,0], nonorth=true)

Constructor for tb_kspace

TightlyBound.TB.myfft — Function

function myfft(crys, nonorth, grid, kpts,ham_kS, Sk=missing)

Does Fourier Transform K->R (ifft) using FFTW.

Arguments

crys crystal

nonorth nonorogonal bool

grid k-point grid size

kpts the k-points nkpts×3 in the original order, to be rearranged into grid

ham_kS hamiltonian in k space (nw×nw×nkpts)

Sk overlaps in k space

TightlyBound.TB.myfft_R_to_K — Function

function myfft_R_to_K(tbc, grid=missing)

Does Fourier Transform R->k (fft) using FFTW for tb_crys

TightlyBound.TB.organizedata — Method

function organizedata(crys::crystal, h::tb)

TightlyBound.TB.organizedata — Method

function organizedata(tbc::tb_crys)

Rearrange the data in the tbc as a function of distances for plotting purposes.

return data_onsite, data_arr

Returns two arrays. The first has data on the onsite elements.

column 1 and 3 have atom indexes

column 2 and 4 have orbital numbers

columns 5 and 6 have real and imaginary parts of H

columns 11 and 12 have real and imaginary parts of S

column 7 has the closest inter-atomic distance

column 8 has the index of the closest atom.

The second has intersite data

column 1 and 3 have atom indexes for the atom pairs

column 2 and 4 have orbital numbers

columns 5 and 6 have real and imaginary parts of H

columns 11 and 12 have real and imaginary parts of S

column 7 has the inter-atomic distance

column 8,9,10 have the the direction cosines lmn for the atom pair.

TightlyBound.TB.read_tb_crys — Method

function read_tb_crys(filename, tbc::tb_crys)

Reads and returns from filename a tb_crys object. See write_tb_crys

If cannot find "filename", will look for "filename.xml", "filename.gz", "filename.xml.gz"

Can read gzipped files directly.

TightlyBound.TB.read_tb_crys_kspace — Method

function read_tb_crys_kspace(filename; directory=missing)

Reads and returns from filename a tb_crys_kspace object. See write_tb_crys_kspace

If cannot find "filename", will look for "filename.xml", "filename.gz", "filename.xml.gz"

Can read gzipped files directly.

TightlyBound.TB.renormalize_tb — Method

function renormalize_tb(d::dftout, h::tb)

Shift eigenvalues from DFT calculation so that the band energy matches the DFT total energy.

TightlyBound.TB.summarize_orb — Method

function summarize_orb(orb::Symbol)

Convert exact orbital (:px) to type of orbital :p .

TightlyBound.TB.symm_by_orbitals — Method

function symm_by_orbitals(crys::crystal, mat)

Helper function to re-symmeterize the overlap matrix properly, starting from incomplete k-points.

TightlyBound.TB.tb_indexes — Method

function tb_indexes(d::dftout)

Figures out mapping between DFT projected hamiltonian orbitals and crystal and the wannier orbitals we want.

return wan, semicore, nwan, nsemi, wan_atom, atom_wan

wan has the indexes of the wannier orbitals

semicore has the indexes of semicore states.

nwan number of wannier orbs

nsemi number of semicore states

wan_atom dictionary wannier to atom numbers

atom_wan dictionary atom numbers to wannier orbitals

TightlyBound.TB.trim — Function

function trim(h::tb, tol=0.0002)

Remove terms in tb with abs value smaller than tol. (Ryd) Will speed calculations at cost of accuracy. USE WITH CARE. Usually not necessary except for very large or very detailed calculations.

TightlyBound.TB.types_energy — Method

function types_energy(types)

TightlyBound.TB.types_energy — Method

function types_energy(c::crystal)

TightlyBound.TB.types_energy — Method

function types_energy(tbc::tb_crys)

TightlyBound.TB.types_energy — Method

function types_energy(tbc::tb_crys)

Calculate the reference energy of each atom in crystal. This results in the total energy being indexed to seperated non-spin-polarized atoms. This is arbirary.

TightlyBound.TB.write_tb_crys — Method

function write_tb_crys(filename, tbc::tb_crys)

Writes to filename a tb_crys object, using xml formatting. See read_tb_crys

TightlyBound.TB.write_tb_crys_kspace — Method

function write_tb_crys_kspace(filename, tbc::tb_crys_kspace)

Save a tb_crys_kspace object to xml format. See read_tb_crys_kspace

BandStruct

TightlyBound.BandStruct.get_kpath — Function

function get_kpath(kpath=[0.5 0 0 ; 0 0 0; 0.5 0.5 0.5], names = missing, npts=30)

Construct a k_path for a band structure calculations. Very simple.

kpath high symmetry k-points in fractional BZ coordinates.

names names of kpoints like ["Γ", "X"]

npts number of points between high-symmetry kpoints

TightlyBound.BandStruct.plot_bandstr — Method

function plot_bandstr(h::tb_crys; kpath, names = missing, proj_types=missing, proj_orbs = missing, proj_nums=missing)

Plot the band structure of a tb_crys object. Can also perform a projected band structure if you specify at least one of proj_types, proj_orbs, proj_nums.

k-path specified by a kpath array and names.

Arguments

h::tb_crys - The tight-biding object we want to plot bands from. Only required argument.

kpath=[0.5 0 0 ; 0 0 0; 0.5 0.5 0.5; 0 0.5 0.5; 0 0 0 ;0 0 0.5] - nk × 3 array k-point path (high symmetry points).

npts=30, - number of points between high-symmetry k-points.

names=missing - nk string array. Names of the high-symmetry k-points

proj_types=missing - types to project onto. Either proj_types="H" or proj_types=["H", "O"] are valid.

proj_orbs=missing - orbitals to project onto. either proj_orbs=:s or proj_orbs=[:s, :p].

proj_nums=missing - atom numbers to project onto. Either proj_nums=1 or proj_nums=[1, 2]

efermi=missing - allows you to specify fermi energy. Default is to take from h

color="blue" - specify line color

MarkerSize=missing" - specify markersize

yrange=missing" - specify y-range. e.g. yrange=[-0.7, 0.3]

plot_hk=false - plot things besides the normal band structure. Can be one of :Seig, :Heig, :Hreal, :Himag, :Sreal, :Simag to plot H or S eigvals or components. Primarily for debugging.

align="vbm" - default or "valence" is to align valence band max to zero energy. Can also be "min", which aligns on the minimum eigenvalue, or "fermi" or "ef", which align on the Fermi level,

clear_pervious=true - clears the plot before adding new stuff.

do_display=true - display the plot. If false, can be used with display-less nodes. You can still use savefig from Plots to produce saved images.

TightlyBound.BandStruct.plot_bandstr — Method

function plot_bandstr(h::tb)

Plots using tb

TightlyBound.BandStruct.plot_compare_dft — Method

function plot_compare_dft(tbc::tb_crys, bs; tbc2=missing)

Plots a band structure comparison between a tight-binding crystal object (tb_crys) and a band structure directly from dft (either a dftout or bs object).

The k-points are fixed by the bs object.

tbc2 is an optional second tbc_crys.

TightlyBound.BandStruct.plot_compare_tb — Method

function plot_compare_tb(h1::tb, h2::tb; h3=missing)

TightlyBound.BandStruct.plot_compare_tb — Method

function plot_compare_tb(h1::tb_crys, h2::tb_crys; h3=missing)

Plot a comparison between different tight binding objects h1, h2, and optionally h3. Options similar to plot_bandstr but more limited.

DOS

TightlyBound.DOS.dos — Method

function dos(tbc::tb_crys; grid=missing, npts=300, proj_type=missing, do_display=true)

DOS, using tetrahedral integration

grid is the k-point grid. Defaults to 1.6 times the default grid for energy integration.

npts is number of energies

proj_type can be "none", "atomic", or "orbital". Defaults to atomic if more than one atom type.

do_display=false will suppress plotting

return energies, dos, projecteddos, pdosnames

TightlyBound.DOS.gaussian_dos — Method

function gaussian_dos(tbc::tb_crys; grid=missing, smearing=0.02, npts=500, proj_type=missing, do_display=true)

Simple Gaussian DOS, mostly for testing.

npts is number of energies

proj_type can be "none", "atomic", or "orbital"

do_display=false will suppress the actual plot

The combination of smearing and grid are important to get converged results.

See also dos

return energies, dos, projecteddos, pdosnames

TightlyBound.DOS.plot_dos — Method

function plot_dos(energies, dos, pdos, names; filename=missing, do_display=true)

Does the actual DOS plotting, called by dos or gaussian_dos

TightlyBound.DOS.projection — Method

function projection(tbc::tb_crys, vects, sk3, grid; ptype=missing)

Figures out the projections.

TightlyBound.DOS.setup_tetra — Method

function setup_tetra(grid)

Setup simple tetrahedron method based on tetra.f90 in QE. Simple tetra method.