TightlyBound

TB

TightlyBound.TB.tb_k — Type

mutable struct tb_k{T}

Tight binding object in k-space. Can be from direct import of DFT band structure using atomic proj, or from fft'ed tb object. Similar to real-space version.

Holds

Hk::Array{Complex{T},3} Hamiltonian in k-space

K::Array{Float64,2} K point array. In fractional coordinates of BZ.

kweights::Array{Float64,1} K point weights.

k_dict::Dict Dictionary from k-point like [0,0,0] to index.

nwan::Int64 Number of orbitals / generalized wannier functions.

nk::Int64 Number of k-points.

nonorth::Bool

Sk::Array{Complex{T},3} Overlap matrix.

scf::Bool needs self-consistency?

h1::Array{T,2} #scf term holds scf term if present

grid::Array{Int64,1} dimensions of k-point grid, from regular grid like [8,8,8]

TightlyBound.TB.tb_crys_kspace — Type

mutable struct tb_crys_kspace{T}

Hold k-point tight binding and crystal structure. Similar to tb_crys

Holds

tb::tb_k

crys::crystal

nelec::Float64

dftenergy::Float64

scf::Bool

gamma::Array{T, 2}

eden::Array{Float64,1}

TightlyBound.TB.read_tb_crys_kspace — Function

function read_tb_crys_kspace(filename; directory=missing)

Reads and returns from filename a tb_crys_kspace object. See write_tb_crys_kspace

If cannot find "filename", will look for "filename.xml", "filename.gz", "filename.xml.gz"

Can read gzipped files directly.

TightlyBound.TB.write_tb_crys_kspace — Function

function write_tb_crys_kspace(filename, tbc::tb_crys_kspace)

Save a tb_crys_kspace object to xml format. See read_tb_crys_kspace

TightlyBound.TB.make_tb_crys — Function

function make_tb_crys(ham::tb,crys::crystal, nelec::Float64, dftenergy::Float64; scf=false, eden = missing, gamma=missing, within_fit=true, screening=1.0, tb_energy=-999, fermi_energy=0.0 )

Constructor function for tb_crys object

TightlyBound.TB.make_tb_crys_kspace — Function

function make_tb_crys_kspace(hamk::tb_k,crys::crystal, nelec::Float64, dftenergy::Float64; scf=false, eden = missing, gamma=missing, screening=1.0)

Constructor function for tb_crys_kspace object

TightlyBound.TB.make_tb — Function

function make_tb(H, ind_arr, r_dict::Dict; h1=missing)

Constructor function for tb

function make_tb(H, ind_arr, r_dict::Dict, S; h1=missing)

Constructor function for tb with overlaps

function make_tb(H, ind_arr, S; h1=missing)

Constructor function for tb, better programming.

TightlyBound.TB.make_tb_k — Function

function make_tb_k(Hk, K, kweights, Sk; h1=missing, grid=[0,0,0], nonorth=true)

Constructor for tb_kspace

TightlyBound.TB.load_hr_dat — Function

function load_hr_dat(filename, directory="")

Load a wannier90 hr.dat file Not currently a major part of program, but you can use if you want.

TightlyBound.TB.calc_energy_fft — Function

function calc_energy_fft(tbc::tb_crys; grid=missing, smearing=0.01, return_more_info=false)

Get energy using fft.

returns energy

if return_more_info==true then returns etot, efermi, vals, vects

TightlyBound.TB.calc_energy_charge_fft — Function

function calc_energy_charge_fft(tbc::tb_crys; grid=missing, smearing=0.01)

Do fft, then calculate energy and charge.

TightlyBound.TB.calc_energy — Function

function calc_energy(tbc::tb_crys; smearing=0.01, returnk=false)

Calculate energy without fft.

function calc_energy(h::tb_crys, kgrid; smearing=0.01, returnk=false)

Calculate energy no fft

TightlyBound.TB.types_energy — Function

function types_energy(tbc::tb_crys)

Calculate the reference energy of each atom in crystal. This results in the total energy being indexed to seperated non-spin-polarized atoms. This is arbirary.

function types_energy(tbc::tb_crys)

function types_energy(c::crystal)

function types_energy(types)

TightlyBound.TB.make_kgrid — Function

function make_kgrid(kgrid)

-kgrid is an array of 3 integers like [8,8,8]

returns regular MP Gamma-centered k-point grid and (equal) k-weights.

TightlyBound.TB.trim — Function

function trim(h::tb, tol=0.0002)

Remove terms in tb with abs value smaller than tol. (Ryd) Will speed calculations at cost of accuracy. USE WITH CARE. Usually not necessary except for very large or very detailed calculations.

TightlyBound.TB.organizedata — Function

function organizedata(tbc::tb_crys)

Rearrange the data in the tbc as a function of distances for plotting purposes.

return data_onsite, data_arr

Returns two arrays. The first has data on the onsite elements.

column 1 and 3 have atom indexes

column 2 and 4 have orbital numbers

columns 5 and 6 have real and imaginary parts of H

columns 11 and 12 have real and imaginary parts of S

column 7 has the closest inter-atomic distance

column 8 has the index of the closest atom.

The second has intersite data

column 1 and 3 have atom indexes for the atom pairs

column 2 and 4 have orbital numbers

columns 5 and 6 have real and imaginary parts of H

columns 11 and 12 have real and imaginary parts of S

column 7 has the inter-atomic distance

column 8,9,10 have the the direction cosines lmn for the atom pair.

function organizedata(crys::crystal, h::tb)

TightlyBound.TB.myfft_R_to_K — Function

function myfft_R_to_K(tbc, grid=missing)

Does Fourier Transform R->k (fft) using FFTW for tb_crys

TightlyBound.TB.myfft — Function

function myfft(crys, nonorth, grid, kpts,ham_kS, Sk=missing)

Does Fourier Transform K->R (ifft) using FFTW.

Arguments

crys crystal

nonorth nonorogonal bool

grid k-point grid size

kpts the k-points nkpts×3 in the original order, to be rearranged into grid

ham_kS hamiltonian in k space (nw×nw×nkpts)

Sk overlaps in k space

TightlyBound.TB.get_sym_R — Function

function get_sym_R(crys, grid, sss = 1.0)

Figures out the r-space grid using Wigner-Seitz like construction to figure out the best arrangement of r-grid points to keep periodic copies closest to the original atom and take into account symmetry.

return R_grid, R_int_grid, sym_R

returns the R_grid, the integer version, and the symmetry factor of each point.

TightlyBound.TB.tb_indexes — Function

function tb_indexes(d::dftout)

Figures out mapping between DFT projected hamiltonian orbitals and crystal and the wannier orbitals we want.

return wan, semicore, nwan, nsemi, wan_atom, atom_wan

wan has the indexes of the wannier orbitals

semicore has the indexes of semicore states.

nwan number of wannier orbs

nsemi number of semicore states

wan_atom dictionary wannier to atom numbers

atom_wan dictionary atom numbers to wannier orbitals

TightlyBound.TB.find_vbm_cbm — Function

function find_vbm_cbm(eigs, fermi)

Find the valence band max and conduction band minimum from eigs, relative to Fermi level.

TightlyBound.TB.ewald_energy — Function

function ewald_energy(tbc::tb_crys, delta_q=missing)

Return ewald energy term from tbc. If delta_q, the atomic charge density, is missing, loads from tbc.

function ewald_energy(tbc::tb_crys_kspace, delta_q=missing)

function ewald_energy(crys::crystal, gamma, delta_q::Array{Float64,1})

Does the actual calculation.

TightlyBound.TB.get_dq — Function

function get_dq(tbc::tb_crys_kspace)

Get atomic charge density from tb_crys or tb_crys_kspace or crys + eden

function get_dq(tbc::tb_crys)

function get_dq(crys::crystal, chargeden::Array{Float64,1})

TightlyBound.TB.get_energy_electron_density_kspace — Function

function get_energy_electron_density_kspace(tbcK::tb_crys_kspace; smearing = 0.01)

Get energy / charge density from k-space tight binding object.

return bandenergy + etypes + echarge + energy_smear, eden, VECTS, VALS, error_flag

function get_energy_electron_density_kspace(tb_k::tb_k, nelec; smearing = 0.01)

K-space get energy and electron density from tb_k

BandStruct

TightlyBound.BandStruct.get_kpath — Function

function get_kpath(kpath=[0.5 0 0 ; 0 0 0; 0.5 0.5 0.5], names = missing, npts=30)

Construct a k_path for a band structure calculations. Very simple.

kpath high symmetry k-points in fractional BZ coordinates.

names names of kpoints like ["Γ", "X"]

npts number of points between high-symmetry kpoints

TightlyBound.BandStruct.plot_bandstr — Method

function plot_bandstr(h::tb_crys; kpath, names = missing, proj_types=missing, proj_orbs = missing, proj_nums=missing)

Plot the band structure of a tb_crys object. Can also perform a projected band structure if you specify at least one of proj_types, proj_orbs, proj_nums.

k-path specified by a kpath array and names.

Must do scf calculation before plotting.

Arguments

h::tb_crys - The tight-biding object we want to plot bands from. Only required argument.

kpath=[0.5 0 0 ; 0 0 0; 0.5 0.5 0.5; 0 0.5 0.5; 0 0 0 ;0 0 0.5] - nk × 3 array k-point path (high symmetry points).

npts=30, - number of points between high-symmetry k-points.

names=missing - nk string array. Names of the high-symmetry k-points

proj_types=missing - types to project onto. Either proj_types="H" or proj_types=["H", "O"] are valid.

proj_orbs=missing - orbitals to project onto. either proj_orbs=:s or proj_orbs=[:s, :p].

proj_nums=missing - atom numbers to project onto. Either proj_nums=1 or proj_nums=[1, 2]

efermi=missing - allows you to specify fermi energy. Default is to take from h

color="blue" - specify line color

MarkerSize=missing" - specify markersize

yrange=missing" - specify y-range. e.g. yrange=[-0.7, 0.3]

plot_hk=false - plot things besides the normal band structure. Can be one of :Seig, :Heig, :Hreal, :Himag, :Sreal, :Simag to plot H or S eigvals or components. Primarily for debugging.

align="vbm" - default or "valence" is to align valence band max to zero energy. Can also be "min", which aligns on the minimum eigenvalue, or "fermi" or "ef", which align on the Fermi level,

clear_pervious=true - clears the plot before adding new stuff.

do_display=true - display the plot. If false, can be used with display-less nodes. You can still use savefig from Plots to produce saved images.

TightlyBound.BandStruct.plot_bandstr — Method

function plot_bandstr(h::tb)

Plots using tb

TightlyBound.BandStruct.plot_compare_dft — Method

function plot_compare_dft(tbc::tb_crys, bs; tbc2=missing)

Plots a band structure comparison between a tight-binding crystal object (tb_crys) and a band structure directly from dft (either a dftout or bs object).

The k-points are fixed by the bs object.

tbc2 is an optional second tbc_crys.

TightlyBound.BandStruct.plot_compare_tb — Method

function plot_compare_tb(h1::tb, h2::tb; h3=missing)

TightlyBound.BandStruct.plot_compare_tb — Method

function plot_compare_tb(h1::tb_crys, h2::tb_crys; h3=missing)

Plot a comparison between different tight binding objects h1, h2, and optionally h3. Options similar to plot_bandstr but more limited.

DOS

TightlyBound.DOS.dos — Method

function dos(tbc::tb_crys; grid=missing, npts=missing, proj_type=missing, do_display=true)

DOS, using tetrahedral integration

grid is the k-point grid. Defaults to 1.6 times the default grid for energy integration.

npts is number of energies

proj_type can be "none", "atomic", or "orbital". Defaults to atomic if more than one atom type.

do_display=false will suppress plotting

return energies, dos, projected_dos, pdos_names

TightlyBound.DOS.gaussian_dos — Method

function gaussian_dos(tbc::tb_crys; grid=missing, smearing=0.02, npts=missing, proj_type=missing, do_display=true)

Simple Gaussian DOS, mostly for testing.

npts is number of energies

proj_type can be "none", "atomic", or "orbital"

do_display=false will suppress the actual plot

The combination of smearing and grid are important to get converged results.

See also dos

return energies, dos, projected_dos, pdos_names

TightlyBound.DOS.plot_dos — Method

function plot_dos(energies, dos, pdos, names; filename=missing, do_display=true)

Does the actual DOS plotting, called by dos or gaussian_dos

TightlyBound.DOS.projection — Method

function projection(tbc::tb_crys, vects, sk3, grid; ptype=missing)

Figures out the projections. ptype can be :atomic or :orbs for atom projection or orbital projection (:s,:p,:d) Default is to choose :atomic except for elemental systems.

TightlyBound.DOS.setup_tetra — Method

function setup_tetra(grid)

Setup simple tetrahedron method based on tetra.f90 in QE. Simple tetra method.

Atomdata

TightlyBound.Atomdata — Module

module Atomdata

Holds the data for isolated atoms.

TightlyBound.Atomdata.atom_prefered_oxidation — Constant

atom_prefered_oxidation::Dict()

Prefered oxidation states of atoms, in descending order of preference (approximate).

TightlyBound.Atomdata.atom_radius — Constant

atom_radius::Dict()

Atomic radius (metallic), in pm

TightlyBound.Atomdata.atoms — Constant

atoms::Dict()

Periodic table information.

TightlyBound.Atomdata.cutoff_dist — Constant

cutoff_dist::Dict()

Cutoff distance for Hamiltonian calculation

TightlyBound.Atomdata.electronegativity — Constant

electronegativity::Dict()

Pauling scale.

TightlyBound.Atomdata.min_dimer_dist_dict — Constant

min_dimer_dist_dict::Dict()

Approximate minimum distance in current fitting data.

TightlyBound.Atomdata.sub_list — Constant

sub_list::Dict()

Most likely substitution list, approximate.

AtomicMod

TightlyBound.AtomicMod — Module

module AtomicMod

Simple module for atom type.

TightlyBound.AtomicMod.atom — Type

struct atom

Hold basic atomic information

name::String

Z::Int64 Atomic number

row::Float64 Periodic table row

col::Float64 Periodic table col

mass::Float64 Mass in amu

nval::Float64 Number of valence electrons in TB calculation.

nsemicore::Int64 Number of semicore electrons. Depends on pseudopotential choice

nwan::Int64 Number of TB orbitals

orbitals::Array{Symbol,1} Names of the orbitals, like [:s, :p]

total_energy::Float64 DFT total energy, depends on pseudopotentials.

eigs::Dict orbital eigenvalues of isolated non-spin-polarized nuetral atom.

energy_offset::Float64 energy to add to TB calculation to make isolated atoms have zero energy.

U::Float64 U value for Ewald correction

TightlyBound.AtomicMod.makeatom — Function

function makeatom(name, Z, row, col, mass, nval, nsemicore, orbitals, etot, eigs, vac_potential=0.0, U=0.0)

Constructor for atom.

BandTools

TightlyBound.BandTools — Module

module BandTools

Utility functions for manipulating band structures. These need to be defined early in the code so other Modules have acccess to them.

TightlyBound.BandTools.band_energy — Function

function band_energy(eigs, weights, nelec, smearing = 0.01; returnk=false, returnocc=false, returnef=false, returnboth=false)

Calculate band energy. Has options for additional return variables. Calculates fermi energy internally.

TightlyBound.BandTools.calc_fermi — Function

function calc_fermi(eigs, weights, nelec, smearing = 0.01)

calculate fermi energy using bisection

TightlyBound.BandTools.smearing_energy — Function

function smearing_energy(eigs, weights, efermi, smearing = 0.01)

Smearing contribution to total energy. If you don't take this int account in metals or small gap semiconductors, your energy is not variational in the correct way, and things like forces and stresses become wrong.

CalcTB

TightlyBound.CalcTB — Module

module CalcTB

Create TB matrix sets from coefficients, or prepare to fit the coefficients.

TightlyBound.CalcTB.coefs — Type

struct coefs

Hold the TB coefficients for 2body or 3body interactions

dim::Int64 - dimension (2 or 3)

datH::Array{Float64,1} Hamiltonian parameters

datS::Array{Float64,1} Overlap parameters, if any

sizeH::Int64

sizeS::Int64

inds::Dict{Array{Symbol}, Array{Int64,1}} Holds inds that tell which coeffiecients correspond to which atoms/orbitals

names::Set Names of atoms

orbs::Array{Any,1} Orbitals

cutoff::Float64 cutoff distance

min_dist::Float64 minimum atom-atom distance in fitting data

maxmin_val_train::Dict max and min value of matrix elements within fitting data

dist_frontier::Dict dictionary of pareto frontier of shortest fitting distances

TightlyBound.CalcTB.calc_frontier — Method

function calc_frontier(crys::crystal, frontier; var_type=Float64, test_frontier=false, diststuff=missing, verbose=true)

Calculate a pareto frontier of short distances. For 2body interactions, this is just a single distance, which is easy. For 3body interactions, there are 3 distances, so multiple points are on the frontier. Useful for deciding if old fitting data applies to a new structure with atoms that are close together.

test_frontier=true is used to check if new structure is in the frontier. othersise, see if new structure changes frontier.

TightlyBound.CalcTB.calc_onsite — Function

function calc_onsite(t1,s1,s2, database=missing)

Handles atomic matrix els. We do not currently fit onsite matrix els

TightlyBound.CalcTB.calc_tb_fast — Function

calc_tb_fast(crys::crystal; database=missing, use_threebody=true, use_threebody_onsite=true)

Construct tb_crys from crystal stucture, but does not solve. This is usually called internally by functions like scf_energy, but you can use it directly if you want. Until you do a SCF energy calculation, the electron density and Fermi level will be wrong.

Arguments

crys::crystal - Required crystal structure

database=missing - Source of coefficients. Will load from default source if not specified.

use_threebody=true - Use three-body off-site interactions. Only turn off for testing purposes.

use_threebody_onsite=true - Use three-body on-site interactions. Only turn off for testing purposes.

verbose=true - set to false for less output.

var_type=missing - variable type of tb_crys. Default is Float64.

TightlyBound.CalcTB.calc_tb_prepare_fast — Method

function calc_tb_prepare_fast(reference_tbc::tb_crys; use_threebody=false, use_threebody_onsite=false)

Take a tbc from DFT and rearrange it for use in fitting code. Basically set it up so that it is ready for a least squares linear fit of coefficients.

return twobody_arrays, threebody_arrays, hvec, svec, Rvec, INDvec, h_onsite, ind_conversion, dmin_types, dmin_types3

Where

twobody_arrays - info for fitting twobody coefficients

threebody_array - info for fitting twobody coefficients

hvec - vector of reference TB matrix els, arranged for fitting

svec - vector of reference overlap matrix els, arranged for fitting

Rvec - displaments

INDvec info on which matrix el goes with which row

h_onsite info for subtracting atomic terms, I think

ind_conversion - dict to convert between place in hamiltonian and overall counter, which removes duplicates.

dmin_types - shortest 2body distances

dmin_types - shortest 3body distances

TightlyBound.CalcTB.calc_threebody — Method

function calc_threebody(c,ind, t1,t2,t3,orb1,orb2,dist,dist31,dist32,lmn12, lmn31,lmn32; database=missing, memory0=missing, memory1=missing, memory2=missing, memoryV=missing, precalc=false, set_maxmin=false)

Calculate 3body intersite hamiltonian interactions.

memory1, etc have preallocated memory. This function is important for performance.

TightlyBound.CalcTB.calc_threebody_onsite — Method

function calc_threebody_onsite(t1,t2,t3,orb1,dist12,dist13,dist23, database; set_maxmin=false, memory=missing)

calculate threebody onsite interactions

TightlyBound.CalcTB.calc_twobody — Method

function calc_twobody(t1,t2,orb1,orb2,dist,lmn, database)

Two body intersite Hamiltonian and overlap matrix els.

TightlyBound.CalcTB.calc_twobody_onsite — Method

function calc_twobody_onsite(t1,t2,orb1,orb2, dist,lmn, database)

Calculate 2body onsite interactions.

TightlyBound.CalcTB.distances_etc — Function

function distances_etc(crys, cutoff, cutoff2=missing)

Old distance function, no threebody only twobody. Code primarly uses distances_etc_3bdy_parallel

TightlyBound.CalcTB.distances_etc_3bdy_parallel — Function

function distances_etc_3bdy_parallel(crys, cutoff=missing, cutoff2=missing; var_type=Float64)

Finds atoms-atom and atom-atom-atom distances that are less than the cutoffs.

return R_keep, R_keep_ab, array_ind3, array_floats3, dist_arr, c_zero, dmin_types, dmin_types3

Where

R_keep has the supercells that are necessary

R_keep_ab has the 2body indexes that are less than the cutoff

array_ind3 has the 3body indexes

array_floats3 has the 3body actual distances and lmn.

dist_arr has the actual 2body distances

c_zero the index of cell [0,0,0], the onsite cell

dmin_types minimum distance between types of atoms.

dmin_types3 minimum 3body distances between types of atoms.

TightlyBound.CalcTB.fit_threebody — Method

function fit_threebody(t1,t2,t3,orb1,orb2,dist,dist31,dist32,lmn12, lmn31,lmn32)

Fit threebody intersite interactions

TightlyBound.CalcTB.fit_threebody_onsite — Method

function fit_threebody_onsite(t1,t2,t3,orb1,dist12,dist13,dist23)

Fit three body onsite interactions.

TightlyBound.CalcTB.fit_twobody — Method

function fit_twobody(orb1,orb2,dist,lmn)

Fit Two body intersite Hamiltonian and overlap matrix els.

TightlyBound.CalcTB.fit_twobody_onsite — Method

function fit_twobody_onsite(t1,t2,orb1,orb2, dist,lmn)

Fit 2body onsite interactions.

TightlyBound.CalcTB.get_data_info — Method

function get_data_info(at_set, dim)

Figure out the arrangement of data in a coefs file.

Loops over various combinations of orbitals and atoms and assigns them places in datH and datS, depending on the terms included in the model and the dimensionaly.

TightlyBound.CalcTB.laguerre — Function

function laguerre(dist, ind=missing; nmax=6, memory=missing)

Calculate laguerre polynomials up to order nmax

TightlyBound.CalcTB.make_coefs — Method

function make_coefs(at_list, dim; datH=missing, datS=missing, cutoff=18.01, min_dist = 3.0, fillzeros=false, maxmin_val_train=missing, dist_frontier=missing)

Constructor for coefs. Can create coefs filled with ones for testing purposes.

See coefs to understand arguments.

TightlyBound.CalcTB.plot_database — Function

function plot_database(database, entry, t=missing)

Plot some data from coefs. Needs to be updated to work with Plots, probably doesn't work now.

TightlyBound.CalcTB.read_coefs — Function

function read_coefs(filename, directory = missing)

Read coefs from filename. Can read gzip directly.

TightlyBound.CalcTB.renormalize_S — Function

function renormalize_S(tbc, database, cutoff=17.99)

Function for changing S but keeping the band structure fixed. Not currently used.

TightlyBound.CalcTB.symmetry_factor — Method

function symmetry_factor(s1,s2,lmn, dat)

All of the spd Slater-Koster matrix elements. dat is preallocated memory.

TightlyBound.CalcTB.symmetry_factor_fit — Method

function symmetry_factor_fit(s1,s2,lmn)

All the Slater-Koster factors, for fitting.

TightlyBound.CalcTB.three_body_H — Function

function two_body_H(dist, ind=missing)

get 3body hamiltonian terms together.

TightlyBound.CalcTB.three_body_O — Function

function two_body_O(dist, ind=missing)

three body onsite.

TightlyBound.CalcTB.trim_dist — Function

function trim_dist(tbc, cutoff=18.0001)

Reduce the atom-atom hamiltonian terms longer than cutoff.

Not used in typical code, but can make tbc run faster / reduce memory.

TightlyBound.CalcTB.two_body_H — Function

function two_body_H(dist, ind=missing)

Two body intersite Hamiltonian.

TightlyBound.CalcTB.two_body_O — Function

function two_body_O(dist, ind=missing)

Two body onsite

TightlyBound.CalcTB.two_body_S — Function

function two_body_H(dist, ind=missing)

Two body intersite overlap.

TightlyBound.CalcTB.write_coefs — Method

function write_coefs(filename, co::coefs; compress=true)

Write coefs to a file. Compress uses gzip. See read_coefs

CrystalMod

TightlyBound.CrystalMod.parsePOSCAR — Function

function parsePOSCAR(lines)

Parse a POSCAR from VASP

Called by makecrys, doesn't need to be called directly.

TightlyBound.CrystalMod.parseQEinput — Function

function parseQEinput(lines)

Parse a quantum espresso inputfile. Can only handle simple cases with explict CELL_PARAMETERS Cannot handle nonzero ibrav. or celldm

Called by makecrys, doesn't need to be called directly.

TightlyBound.CrystalMod.generate_supercell — Function

function generate_supercell(crys, cell)

Generate supercell. cell is [1,1,2], etc.

Note, perfered notation is to use syntax: c * [1,1,2] , where c is a crystal, thus using overloading of the * operator, rather than calling directly.

Missing docstring.

Missing docstring for TightlyBound.CrystalMod.generate_random. Check Documenter's build log for details.

TightlyBound.CrystalMod.write_poscar — Function

function write_poscar(crys, filename)

Write a crystal to a POSCAR.

TightlyBound.CrystalMod.write_efs — Function

function function write_efs(crys, energy, forces, stress, filename)

Write crystal, energy, force, stress to fake quantum espresso output file. This is for testing purposes only.

TightlyBound.CrystalMod.get_grid — Function

function get_grid(c, kden=55.0)

Get a default k-point grid size with kden density.

TightlyBound.CrystalMod.orbital_index — Function

function orbital_index(c::crystal)

Get correspondence between crystal and the TB orbital numbers.

return ind2orb, orb2ind, etotal, nval

ind2orb dictionary which gives [atom_number,atom_type, :orbital_symbol] from TB index integer.

orb2ind dictionary which gives TB index integer from [atom_number,atom_type, :orbital_symbol]

etotal total DFT energy of atoms, for calculation atomization energy.

nval number of valence orbitals.

DFToutMod

TightlyBound.DFToutMod — Module

module DFToutMod

Deal with energy, force, stress, band structure, etc, from a DFT calculation

Generic to different DFT codes.

TightlyBound.DFToutMod.bandstructure — Type

mutable struct bandstructure

Band structure. Has

nbnd::Int Number of bands

nks::Int Number of k-points

nelec::Float64 Number of electrons

efermi::Float64 Fermi energy

kpts::Array{Float64,2} List of k-points, nks × 3 array in BZ crystal units.

kweights::Array{Float64,1} Weights of k-points (nks).

kgrid::Array{Int,1} Equivalent gamma centered Monkhorst-Pack k-grid dimensions, if applies.

eigs::Array{Float64,2} Eigenvalues. nks × nbnd

TightlyBound.DFToutMod.dftout — Type

mutable struct dftout

DFT output struct. Has

crys::crystal crystal structure

energy::Float64 the actual DFT energy, depends on pseudopotentials.

energy_smear::Float64 smearing energy

forces::Array{Float64,2} forces Ryd / a.u.

stress::Array{Float64,2} stress Ryd / (a.u.)^3

bandstruct::bandstructure See bandstructure struct

hasband::Bool does this object have a band structure, usually true

hasham::Bool not used, always false

prefix::String A string has a name used to find output files.

outdir::String Directly loaded from

tot_charge::Float64 If charge of unit cell is nonzero

atomize_energy::Float64 Atomization energy, relative to non-spin-polarized atoms.

TightlyBound.DFToutMod.get_atomize_energy — Method

function get_atomize_energy(d::dftout)

Return atomization energy in current energy units

TightlyBound.DFToutMod.makebs — Method

function makebs(nelec::Number, efermi::Number,  kpoints, kweights,kgrid, vals)

Constructor for bandstructure.

TightlyBound.DFToutMod.makedftout — Function

function makedftout(A, pos, types, energy::Number,energy_smear::Number,  forces, stress, bandstruct=missing; prefix="PREFIX", outdir="TMPDIR", tot_charge=0.0)

TightlyBound.DFToutMod.makedftout — Function

function makedftout(crys::crystal, energy::Number, energy_smear::Number,  forces, stress, bandstruct=missing; prefix="PREFIX", outdir="TMPDIR", tot_charge=0.0)

Constructor for dftout. Usually called by function that loads DFT output files, not called directly.

DFT

TightlyBound.DFT — Module

module DFT

This is the generic DFT interface. Only QE is currently implemented however.

TightlyBound.DFT.runSCF — Method

function runSCF(crys::crystal; inputstr=missing, prefix=missing, tmpdir="./", directory="./", functional="PBESOL", wannier=0, nprocs=1, code="QE", skip=false, calculation="scf", dofree="all", tot_charge = 0.0, smearing = missing, magnetic=false, cleanup=false, use_backup=false)

Workflow for generic DFT SCF calculation. code can only by "QE"

QE

TightlyBound.QE — Module

module QE

Module for running Quantum Espresso. Generic DFT version below.

TightlyBound.QE.doclean — Method

function doclean(d)

Clean up wavefunctions in directory d

TightlyBound.QE.loadXML — Method

function loadXML(savedir)

Load a QE SCF DFT calculation into a dftout object.

TightlyBound.QE.loadXML_bs — Method

function loadXML_bs(savedir::String)

Load bandstructure from QE xml file that was converted to a dict already

TightlyBound.QE.loadXML_bs — Method

function loadXML_bs(savedir::String)

Load bandstructure from QE xml file.

TightlyBound.QE.makeSCF — Function

function makeSCF(crys::crystal, directory="./", prefix=missing, tmpdir=missing, functional="PBESOL", wannier=0, calculation="scf", dofree="all", tot_charge = 0.0, smearing = 0.01, magnetic=false; mixing="local-TF")

Make QE inputfile for SCF DFT calculation.

TightlyBound.QE.make_commands — Function

function make_commands(nprocs=1)

Returns a dictionary with command lines to call external programs on the command line nprocs is the number of processors for parallel execution Needs to be changed for you specfic program locations and mpi commands (if any)

This needs to be edited to actually run QE yourself. Running wannier90 is optional, not part of current code.

TightlyBound.QE.makedict — Method

function makedict(savedir)

Load a data-file-schema.xml as julia dictionary.

TightlyBound.QE.runSCF — Function

function runSCF(crys::crystal, inputstr=missing, prefix=missing, tmpdir="./", directory="./", functional="PBESOL", wannier=0, nprocs=1, skip=false, calculation="scf", dofree="all", tot_charge = 0.0, smearing = 0.01, magnetic=false, cleanup=false, use_backup=false)

Workflow for doing SCF DFT calculation on crys

Return dftout

TightlyBound.QE.run_pwscf — Function

function run_pwscf(inputstr, outputstr, nprocs=1, directory="./", use_backup=false)

Run the pw.x code from QE on inputstr

Ewald

TightlyBound.Ewald — Module

module Ewald

Module for electrostatic preperation.

TightlyBound.Ewald.electrostatics_getgamma — Method

function electrostatics_getgamma(crys::crystal;  kappa=missing, noU=false, onlyU=false, screening = 1.0)

Main function. Does Ewald calculation on crys. Returns gamma, which is used in scf calculation. This is only run once for a given tb_crys object and stored.

kappa is the splitting parameter between real/k-space in Ewald calculation. Will estimate best one if not provided.

noU=false for testing only

onlyU=false for testing only

screening=1.0 Not used. Purpose is to reduce U values for values < 1.

TightlyBound.Ewald.estimate_best_kappa — Method

function estimate_best_kappa(A)

Estimate best value of kappa for Ewald sum. Shouldn't effect final value, only calculation speed. There is probably a better way to do this.

TightlyBound.Ewald.getU — Method

function getU(types)

Get values of U from Atomdata.

TightlyBound.Ewald.get_onsite — Method

function get_onsite(crys::crystal, U::Array{Float64,1})

return Onsite terms, Coulumb U.

TightlyBound.Ewald.k_space — Function

function k_space(crys::crystal, kappa, starting_size_kspace=2)

K-space Ewald sum.

TightlyBound.Ewald.real_space — Function

function real_space(crys::crystal, kappa::Float64, U::Array{Float64}, starting_size_rspace=2)

Real-space Ewald sum.

FitTB

TightlyBound.FitTB.do_fitting — Method

function do_fitting(list_of_tbcs; kpoints = missing,  atoms_to_fit=missing, fit_threebody=true, fit_threebody_onsite=true, do_plot = true)

Used for simple linear fitting of coefficients. Interface for more complicated fitting.

list_of_tbcs - List of tbc_crys or tbccrysk` objects to fit to.

dft_list - for kspace fitting, use

fit_threebody=true - Fit threebody coefficients. Sometimes false for testing, but true for production.

fit_threebody_onsite=true - Fit threebody onsite coefficients. See above.

do_plot=true - show simple plot comparing coefficients to tbc reference.

TightlyBound.FitTB.do_fitting_linear — Method

function do_fitting_linear(list_of_tbcs; kpoints = missing, dft_list = missing,  fit_threebody=true, fit_threebody_onsite=true, do_plot = true, starting_database=missing, mode=:kspace, return_database=true, NLIM=100, refit_database=missing)

Linear fitting (not recursive). Used as starting point of recursive fitting.

Arguments.

list_of_tbcs The main data to fit to. Consists of a list of tb_crys or tb_crys_kspace objects.

kpoints = missing Kpoints to do fitting to in k-space. Usually, this is not used, see below.

dft_list = missing List of dftout objects. Normally, we get symmetry reduced k-grids from these objects.

fit_threebody=true Fit three body coefficients. Yes for production runs.

fit_threebody_onsite=true Fit 3body onsite coefficients. Yes for production runs.

do_plot = true Make a plot to assess fitting.

starting_database=missing Use a database dict with some of the coefficents already fit and fixed.

mode=:kspace Fit in either :kspace or :rspace. Can only use r-space if using only tbc_crys real-space objects to fit to. :kspace is normal.

return_database=true Return the final database. For use when called by other functions.

NLIM=100 Largest number of k-points per structure. Set to smaller numbers to make code go faster / reduce memory, but may be less accurate.

refit_database=missing starting point for coefficients we are fitting. Usually not used, as it doesn't always speed things up in practice. Something may not work about this option.

TightlyBound.FitTB.do_fitting_recursive — Method

function do_fitting_recursive(list_of_tbcs ; weights_list = missing, dft_list=missing, X_cv = missing, kpoints = [0 0 0; 0 0 0.5; 0 0.5 0.5; 0.5 0.5 0.5], starting_database = missing,  update_all = false, fit_threebody=true, fit_threebody_onsite=true, do_plot = false, energy_weight = missing, rs_weight=missing,ks_weight=missing, niters=50, lambda=0.0, leave_one_out=false, prepare_data = missing, RW_PARAM=0.0, NLIM = 100, refit_database = missing, start_small = false)

This is the primary function for fitting. Uses the self-consistent linear fitting algorithm.

Arguments

list_of_tbcs List of tbc_crys or tbc_crys_kspace object to fit to.

weights_list = missing relative weights of different tbc objects in fitting code.

dft_list=missing List of dftout objects used to get symmetry-reduced kpoint lists / weights to use in kspace fitting.

kpoints = [0 0 0; 0 0 0.5; 0 0.5 0.5; 0.5 0.5 0.5] alternate way of setting k-points, not usually used.

starting_database = missing If using already fit coefficents for some of the atoms, from another calculation, include that database dict here.

update_all = false If updateall is true, then we refit the starting coefficients from `startingdatabase. Normallyfalse` except for testing.

fit_threebody=true Fit 3body intersite coefficents. true for production runs.

fit_threebody_onsite=true Fit 3body onsite coefficents. true for production runs.

do_plot = false make plot for the linear fit.

energy_weight = missing Weighting for the total energy terms in the fit.

rs_weight=missing Real space hamiltonian matrix els weighting. zero for pure k-space fit.

ks_weight=missing K-space hamiltonian matrix els weighting. Set to zero to ignore hamiltonian matrix els and only fit to band structure.

niters=50 Maximum number of iterations.

lambda=0.0 If greater than zero, include a simple ridge regression with this lambda value. Usually zero.

leave_one_out=false Leave-one-out cross-validation. Too slow to be very useful.

prepare_data = missing Rarely used option to reuse previous linear fitting.

RW_PARAM=0.0 Weighting of non-occupied bands in fit.

NLIM = 100 Maximum k-points per structure. Smaller for faster but less accurate fit that uses less memory.

refit_database = missing Option to include starting data. Rarely used.

start_small = false When fitting only 3body data, setting this to true will start the 3body terms with very small values, which can improve convergence. Not useful if also fitting 2body terms.

TightlyBound.FitTB.extract_database — Method

function extract_database(database_old,nh,ns, KEYS, HIND, SIND)

If we are using fixed prefit coefficients, we have to get them in the same form as our current fitting.

TightlyBound.FitTB.fourierspace — Method

function fourierspace(tbc, kpoints, X_H, X_S, Y_H, Y_S, Xhc, Xsc, rind, Rvec, INDVec, h_on, ind_convert)

Do analytic fourier transform of real space fitting matrices into kspace.

TightlyBound.FitTB.get_k — Method

function get_k(dft_list, ncalc; NLIM = 100)

Decide which k-points to include in fitting, as we limit the total number to NLIM or less per structure

Uses some randomness, but puts high symmetry points at front of line.

TightlyBound.FitTB.hermetian_index — Method

function hermetian_index(i::Int64,j::Int64,nwan::Int64)

This is used to reduce memory by only keeping track of independet coefficients of Hermetian matrices, which is nearly a factor of 2 reduction.

TightlyBound.FitTB.make_database — Method

function make_database(ch, cs,  KEYS, HIND, SIND, DMIN_TYPES, DMIN_TYPES3; scf=false, starting_database=missing, tbc_list=missing)

Construct the coefs and database from final results of fitting.

TightlyBound.FitTB.prepare_for_fitting — Method

function prepare_for_fitting(list_of_tbcs; kpoints = missing, dft_list = missing, fit_threebody=false, fit_threebody_onsite=false, starting_database=missing, refit_database=missing)

Make lots of preperations for fitting. Moves things around, put stuff in materices, etc.

Force_Stress

TightlyBound.Force_Stress — Module

module Force_Stress

Module for calculating force and stress

TightlyBound.Force_Stress.finite_diff — Method

function finite_diff(crys::crystal, database, ind1, ind2; stress_mode=false, step = 0.0002, smearing = 0.01, grid = missing)

Finite differences force/stress, for testing.

Arguments

crys::crystal Crystal structure

database Database of fitting coefficents.

ind1 atom index for first stress index

ind2 cartesian index or second stress index

stress_mode=false true for stress, otherwise force.

step = 0.0002 step_size for finite steps.

smearing = 0.01 smearing energy

grid = missing kpoint grid

TightlyBound.Force_Stress.get_energy_force_stress — Method

function get_energy_force_stress(crys::crystal, database; smearing = 0.01, grid = missing)

Get force and stress, non-fft algorithm. Generally use the fft algorithm.

return energy_tot, f_cart, stress

Returns Ryd units. Generally users should use the scf_energy_force_stress function

Uses automatic differentation for gradient.

TightlyBound.Force_Stress.get_energy_force_stress — Method

function get_energy_force_stress(crys::crystal, database; smearing = 0.01, grid = missing)

Get force and stress, non-fft algorithm

TightlyBound.Force_Stress.get_energy_force_stress_fft — Method

function get_energy_force_stress_fft(tbc::tb_crys, database; do_scf=false, smearing = 0.01, grid = missing, e_den0=missing, vv = missing)

Calculate energy/force/stress using fft algorithm. Users should use scf_energy_force_stress, which calls this. Uses automatic differentation for jacobian.

TightlyBound.Force_Stress.inv_reshape_vec — Method

function inv_reshape_vec(x, strain, nat; strain_mode=true)

The force and relax algorithms from outside codes take in vectors, not crystal 's. So we have to reshape vectors to and from crystals

TightlyBound.Force_Stress.psi_gradH_psi — Method

function psi_gradH_psi(VALS0, VECTS, hk_g, sk_g, h1, VALS, scf, nwan, nat, grid, OCCS)

Helper function for <psi | grad_Ham | psi >

TightlyBound.Force_Stress.psi_gradH_psi2 — Method

function psi_gradH_psi(VALS0, VECTS, hk_g, sk_g, h1, VALS, scf, nwan, nat, grid, OCCS)

Helper function for <psi | grad_Ham | psi >

This version isn't used.

TightlyBound.Force_Stress.reshape_vec — Method

function reshape_vec(x, nat; strain_mode=false)

The force and relax algorithms from outside codes take in vectors, not crystal 's. So we have to reshape vectors to and from crystals

TightlyBound.Force_Stress.safe_mode_energy — Method

function safe_mode_energy(crys::crystal, database; var_type=Float64)

Relaxation can accidently lead to very small atom-atom distances during the relaxation precedure if too large of step is taken. This function is a repulsive energy function at short range to make sure the relaxtion doesn't get stuck at very short distances where the fitting doesn't apply.

ManageDatabase

TightlyBound.ManageDatabase — Module

module ManageDatabase

Module for reading coefs from files and making database as needed for calculations

TightlyBound.ManageDatabase.add_to_database — Method

function add_to_database(s::Set)

Load elements or twobody terms from precalcuated coefs from files.

TightlyBound.ManageDatabase.clear_database — Method

function clear_database()

Remove loaded databasea

TightlyBound.ManageDatabase.prepare_database — Method

function prepare_database(at_list)

TightlyBound.ManageDatabase.prepare_database — Method

function prepare_database(c::crystal)

Get ready database of precalculated coefs for crystal

SCF

TightlyBound.SCF — Module

module SCF

Module for self-consistent field calculations for TB objects.

TightlyBound.SCF.remove_scf_from_tbc — Method

function remove_scf_from_tbc(hk3, sk3, tbc; smearing=0.01, e_den = missing)

This function takes a hk3, sk3 set of hamiltonian / overlap that does not require scf and adjusts it so that it does require scf, but gives the same energy and band structure.

TightlyBound.SCF.remove_scf_from_tbc — Method

function remove_scf_from_tbc(tbcK::tb_crys_kspace; smearing=0.01, e_den = missing)

This function takes a tbc_crys_kspace object that does not require scf and adjusts it so that it does require scf, but gives the same energy and band structure.

TightlyBound.SCF.remove_scf_from_tbc — Method

function remove_scf_from_tbc(tbc::tb_crys; smearing=0.01, grid = missing, e_den = missing)

This function takes a tbc_crys object that does not require scf and adjusts it so that it does require scf, but gives the same energy and band structure.

TightlyBound.SCF.scf_energy — Method

function scf_energy(c::crystal, database::Dict; smearing=0.01, grid = missing, conv_thr = 1e-5, iters = 75, mix = -1.0, mixing_mode=:pulay, verbose=true)

Run scf calculation of c::crystal, using database of coefs. The main user version is scf_energy in TightlyBound, which calls this one.

smearing is smearing energy in Ryd.

grid is k-point grid (gamma centered MP), will use default.

conv_thr convergence threshold in Ryd.

iters maximum iterations for first attempt

mix=-1.0 default is choose mixing for you. Otherwise, set between 0.0 and 1.0

mixing_mode=:pulay default using Pulay mixing (DIIS). Any other input uses simple mixing.

verbose=true verbosity level.

TightlyBound.SCF.scf_energy — Method

function scf_energy(tbc::tb_crys; smearing=0.01, grid = missing, e_den0 = missing, conv_thr = 1e-5, iters = 75, mix = -1.0, mixing_mode=:pulay, verbose=true)

Utilty

TightlyBound.Utility — Module

module Utility

Some useful functions, mostly for converting stuff and loading files.